extract biz logic from aws api gateway logic
