+++ backend/onconnect/src/const.ts
@@ -1,0 +1,2 @@
+export const EXPIRY_TIME_IN_HOUR = process.env.EXPIRY_TIME_IN_HOUR || '16';
+export const TABLE_NAME = process.env.TABLE_NAME || 'scrum_poker';
