+++ backend/sendmessage/src/const.ts
@@ -1,0 +1,1 @@
+export const TABLE_NAME = process.env.TABLE_NAME || 'scrum_poker';
